Fix ReScript indexers without shamefully hoisting deps
