+++ codegenerator/cli/npm/envio/.gitignore
@@ -134,4 +134,4 @@ pnpm-lock.yaml
 lib
 
 # ReScript artifacts
-*.bs.js
+*.res.js
