+++ codegenerator/cli/npm/envio/package.json
@@ -24,6 +24,8 @@
     "rescript": "11.1.3",
     "rescript-schema": "9.3.0",
     "viem": "2.21.0",
-    "bignumber.js": "9.1.2"
+    "bignumber.js": "9.1.2",
+    "pino": "8.16.1",
+    "pino-pretty": "10.2.3"
   }
 }
