+++ codegenerator/cli/npm/envio/package.json.tmpl
@@ -35,7 +35,9 @@
     "rescript": "11.1.3",
     "rescript-schema": "9.3.0",
     "viem": "2.21.0",
-    "bignumber.js": "9.1.2"
+    "bignumber.js": "9.1.2",
+    "pino": "8.16.1",
+    "pino-pretty": "10.2.3"
   },
   "files": [
     "bin.js",
@@ -44,7 +46,6 @@
     "rescript.json",
     "index.d.ts",
     "index.js",
-    "src",
-    "!src/**/*.bs.js"
+    "src"
   ]
 }
