+++ codegenerator/cli/npm/envio/rescript.json
@@ -7,7 +7,7 @@
       "subdirs": true
     }
   ],
-  "suffix": ".bs.js",
+  "suffix": ".res.js",
   "package-specs": {
     "module": "commonjs",
     "in-source": true
