+++ codegenerator/cli/src/cli_args/init_config.rs
@@ -340,7 +340,7 @@ impl Language {
     // Logic to get the event handler directory based on the language
     pub fn get_event_handler_directory(&self) -> String {
         match self {
-            Language::ReScript => "./src/EventHandlers.bs.js".to_string(),
+            Language::ReScript => "./src/EventHandlers.res.js".to_string(),
             Language::TypeScript => "src/EventHandlers.ts".to_string(),
             Language::JavaScript => "./src/EventHandlers.js".to_string(),
         }
