+++ codegenerator/cli/templates/static/codegen/index.js
@@ -2,8 +2,8 @@
  This file serves as an entry point when referencing generated as a node module
  */
 
-const handlers = require("./src/Handlers.bs");
-const TestHelpers = require("./src/TestHelpers.bs");
+const handlers = require("./src/Handlers.res.js");
+const TestHelpers = require("./src/TestHelpers.res.js");
 const BigDecimal = require("bignumber.js");
 
 module.exports = {
