+++ codegenerator/cli/templates/static/greeter_template/rescript/rescript.json
@@ -18,6 +18,6 @@
   "jsx": {
     "version": 4
   },
-  "suffix": ".bs.js",
+  "suffix": ".res.js",
   "bs-dependencies": ["generated", "envio"]
 }
