+++ codegenerator/integration_tests/tests/test_indexers/test_exits/.gitignore
@@ -30,7 +30,6 @@ artifacts
 cache
 generated
 logs
-*.bs.js
-*.bs.mjs
+*.res.js
 *.gen.ts
 build
