+++ scenarios/erc20_multichain_factory/.gitignore
@@ -29,6 +29,6 @@ build/
 benchmarks/
 artifacts
 cache
-*.bs.js
+*.res.js
 generated
 .env
